requirements.txt
liboqs-python 
slixmpp 
matplotlib
pandas